Formula for installing ATNF MIRIAD.
This is a tricky one because there's no install target for make. MIRIAD
expects to be extracted and compiled in the same place you intend to
use it, so some fiddling has to be done after the files are moved to
the brew cellar (i.e. change the MIRCH.sh script).
I'm also annoyed that I have to create the symlink myself. Please fix
it if you figure out why brew wasn't doing it.